Add files via upload
The Peltier Refrigerator uses 1 peltier plate powered by a dc supply. The cooling end of peltiers is connected to inner metal frame to pass on cooling to the contents of fridge. The heating end of the peltier is connected to a heatsink using thermal paste to pass on the heat into environment using cooling fan.                                                           
Peltier module - (thermoelectric module) is a thermal control module that has both warming and cooling effects. By passing an electric current through the module, it is possible to change the surface temperature and keep it at the target temperature.                                                      
The Peltier Refrigerator is a portable, self-sustaining, solar-powered, collapsible refrigerator and freezer that can go anywhere.